Mobile Item 95: New Competitions Entry Functionality added in MHS GOLF
 Functionality, if Teamsize = 4 | 3 | 2
 I would suggest we have the following prompt when we click on the x beside a player
 _Do you wish to change this player or remove the player & partner(s)?

 Prompt _'Change Player_'   or  _Remove the Player & Partner(s)_

 If User selects 'Change Player' use existing functionaily.

 If User selects 'Remover Player & Partner(s)  - remove the team

+++ MHS Demo/MHSystems/app/src/main/java/com/mh/systems/demoapp/web/models/competitionsentrynew/AllPlayer.java
@@ -19,7 +19,7 @@ public class AllPlayer implements Serializable {
     private Boolean IsFriend;
     @SerializedName("EntryFee")
     @Expose
-    private Double EntryFee;
+    private float EntryFee;
 
     /**
      * EntryStatus equal to
@@ -57,11 +57,11 @@ public class AllPlayer implements Serializable {
         this.IsFriend = IsFriend;
     }
 
-    public Double getEntryFee() {
+    public float getEntryFee() {
         return EntryFee;
     }
 
-    public void setEntryFee(Double EntryFee) {
+    public void setEntryFee(float EntryFee) {
         this.EntryFee = EntryFee;
     }
